Refactoring src/

+++ src/algorithms/utils/nabla.py
@@ -5,7 +5,7 @@ def nabla(x):
     Applies the finite differences operator associated with tensors of the same shape as x.
     """
     b, c, h, w = x.shape
-    u = torch.zeros((b, c, h, w, 2), device=x.device).type(x.dtype)
+    u = torch.zeros((b, c, h, w, 2), device=x.device, dtype=x.dtype)
     u[:, :, :-1, :, 0] = u[:, :, :-1, :, 0] - x[:, :, :-1]
     u[:, :, :-1, :, 0] = u[:, :, :-1, :, 0] + x[:, :, 1:]
     u[:, :, :, :-1, 1] = u[:, :, :, :-1, 1] - x[..., :-1]
@@ -17,9 +17,8 @@ def nabla_adjoint(x):
     Applies the adjoint of the finite difference operator.
     """
     b, c, h, w = x.shape[:-1]
-    u = torch.zeros((b, c, h, w), device=x.device).type(
-        x.dtype
-    )  # note that we just reversed left and right sides of each line to obtain the transposed operator
+    # note: transposed operator — each line has left/right sides swapped
+    u = torch.zeros((b, c, h, w), device=x.device, dtype=x.dtype)
     u[:, :, :-1] = u[:, :, :-1] - x[:, :, :-1, :, 0]
     u[:, :, 1:] = u[:, :, 1:] + x[:, :, :-1, :, 0]
     u[..., :-1] = u[..., :-1] - x[..., :-1, 1]
